ci: Remove check for github-actions
The tool used for that check doesn't support the full schema of what
github action and workflow yamls can contain. Notably, path globs are
not supported very well. In addition the tool doesn't seem to be
maintained very actively.

Now that the PR required checks are not going to allow merging if Github
refuses to parse the workflow files, let's rely on Github validating and
showing errors in these files. Previously when the required check was
based on poseidon/wait-for-status-checks action, that would had been a
problem as parse error caused job not to start and that action ended up
giving green light for merging.